Fix a few other easy warnings on Mac
Ones that occur a lot because they are in headers
Plus some in SdkTest

+++ include/mega/logging.h
@@ -407,7 +407,7 @@ public:
                     i++;
                 }
 
-                logger->log(nullptr, level, nullptr, "", dm.get(), dms.get(), static_cast<int>(i));
+                logger->log(nullptr, level, nullptr, "", dm.get(), dms.get(), i);
             }
         }
         for (auto &s: mCopiedParts)
